Move MessageReader to pkg/cas
