+++ pkg/blobstore/BUILD.bazel
@@ -35,6 +35,7 @@ go_library(
         "//pkg/blobstore/buffer",
         "//pkg/blobstore/slicing",
         "//pkg/capabilities",
+        "//pkg/cas",
         "//pkg/clock",
         "//pkg/cloud/aws",
         "//pkg/cloud/gcp",
@@ -43,7 +44,6 @@ go_library(
         "//pkg/proto/fsac",
         "//pkg/proto/icas",
         "//pkg/proto/iscc",
-        "//pkg/storage",
         "//pkg/util",
         "//pkg/zstd",
         "@bazel_remote_apis//build/bazel/remote/execution/v2:remote_execution_go_proto",
